test: refactor vote tests
